For the most part, rename whisk to eip7441
