+++ .gitignore
@@ -24,8 +24,8 @@ tests/core/pyspec/eth2spec/capella/
 tests/core/pyspec/eth2spec/deneb/
 tests/core/pyspec/eth2spec/electra/
 tests/core/pyspec/eth2spec/fulu/
-tests/core/pyspec/eth2spec/whisk/
 tests/core/pyspec/eth2spec/eip6800/
+tests/core/pyspec/eth2spec/eip7441/
 tests/core/pyspec/eth2spec/eip7732/
 
 # coverage reports
