+++ configs/mainnet.yaml
@@ -56,9 +56,9 @@ ELECTRA_FORK_EPOCH: 18446744073709551615  # temporary stub
 # Fulu
 FULU_FORK_VERSION: 0x06000000
 FULU_FORK_EPOCH: 18446744073709551615  # temporary stub
-# WHISK
-WHISK_FORK_VERSION: 0x08000000  # temporary stub
-WHISK_FORK_EPOCH: 18446744073709551615
+# EIP7441
+EIP7441_FORK_VERSION: 0x08000000  # temporary stub
+EIP7441_FORK_EPOCH: 18446744073709551615
 # EIP7732
 EIP7732_FORK_VERSION: 0x09000000  # temporary stub
 EIP7732_FORK_EPOCH: 18446744073709551615
@@ -175,10 +175,8 @@ BALANCE_PER_ADDITIONAL_CUSTODY_GROUP: 32000000000
 MAX_BLOBS_PER_BLOCK_FULU: 12
 MIN_EPOCHS_FOR_DATA_COLUMN_SIDECARS_REQUESTS: 4096
 
-# Whisk
-# `Epoch(2**8)`
+# EIP7441
 WHISK_EPOCHS_PER_SHUFFLING_PHASE: 256
-# `Epoch(2)`
 WHISK_PROPOSER_SELECTION_GAP: 2
 
 # EIP7732
