+++ configs/minimal.yaml
@@ -55,9 +55,9 @@ ELECTRA_FORK_EPOCH: 18446744073709551615
 # Fulu
 FULU_FORK_VERSION: 0x06000001
 FULU_FORK_EPOCH: 18446744073709551615
-# WHISK
-WHISK_FORK_VERSION: 0x08000001
-WHISK_FORK_EPOCH: 18446744073709551615
+# EIP7441
+EIP7441_FORK_VERSION: 0x08000001
+EIP7441_FORK_EPOCH: 18446744073709551615
 # EIP7732
 EIP7732_FORK_VERSION: 0x09000001
 EIP7732_FORK_EPOCH: 18446744073709551615
@@ -176,7 +176,7 @@ BALANCE_PER_ADDITIONAL_CUSTODY_GROUP: 32000000000
 MAX_BLOBS_PER_BLOCK_FULU: 12
 MIN_EPOCHS_FOR_DATA_COLUMN_SIDECARS_REQUESTS: 4096
 
-# Whisk
+# EIP7441
 WHISK_EPOCHS_PER_SHUFFLING_PHASE: 4
 WHISK_PROPOSER_SELECTION_GAP: 1
