+++ presets/minimal/eip7441.yaml
@@ -1,15 +1,15 @@
-# Mainnet preset - Whisk
+# Minimal preset - EIP7441
 
 # Misc
 # ---------------------------------------------------------------
-# `uint64(4)`
+# [customized]
 CURDLEPROOFS_N_BLINDERS: 4
-# `uint64(2**14)`
-WHISK_CANDIDATE_TRACKERS_COUNT: 16384
-# `uint64(2**13)` must be < WHISK_CANDIDATE_TRACKERS_COUNT
-WHISK_PROPOSER_TRACKERS_COUNT: 8192
-# `uint64(2**7 - CURDLEPROOFS_N_BLINDERS)`
-WHISK_VALIDATORS_PER_SHUFFLE: 124
+# [customized]
+WHISK_CANDIDATE_TRACKERS_COUNT: 32
+# [customized]
+WHISK_PROPOSER_TRACKERS_COUNT: 16
+# [customized]
+WHISK_VALIDATORS_PER_SHUFFLE: 4
 # `uint64(2**15)` TODO: will be replaced by a fix format once there's a serialized format
 WHISK_MAX_SHUFFLE_PROOF_SIZE: 32768
 # `uint64(2**10)` TODO: will be replaced by a fix format once there's a serialized format
