+++ pysetup/md_doc_paths.py
@@ -8,8 +8,8 @@ from .constants import (
     DENEB,
     ELECTRA,
     FULU,
-    WHISK,
     EIP6800,
+    EIP7441,
     EIP7732,
 )
 
@@ -22,8 +22,8 @@ PREVIOUS_FORK_OF = {
     DENEB: CAPELLA,
     ELECTRA: DENEB,
     FULU: ELECTRA,
-    WHISK: CAPELLA,
     EIP6800: DENEB,
+    EIP7441: CAPELLA,
     EIP7732: ELECTRA,
 }
