+++ pysetup/spec_builders/__init__.py
@@ -5,8 +5,8 @@ from .capella import CapellaSpecBuilder
 from .deneb import DenebSpecBuilder
 from .electra import ElectraSpecBuilder
 from .fulu import FuluSpecBuilder
-from .whisk import WhiskSpecBuilder
 from .eip6800 import EIP6800SpecBuilder
+from .eip7441 import EIP7441SpecBuilder
 from .eip7732 import EIP7732SpecBuilder
 
 
@@ -14,6 +14,6 @@ spec_builders = {
     builder.fork: builder
     for builder in (
         Phase0SpecBuilder, AltairSpecBuilder, BellatrixSpecBuilder, CapellaSpecBuilder, DenebSpecBuilder,
-        ElectraSpecBuilder, FuluSpecBuilder, WhiskSpecBuilder, EIP6800SpecBuilder, EIP7732SpecBuilder,
+        ElectraSpecBuilder, FuluSpecBuilder, EIP6800SpecBuilder, EIP7441SpecBuilder, EIP7732SpecBuilder,
     )
 }
