+++ tests/core/pyspec/eth2spec/test/eip7441/sanity/blocks/__init__.py
@@ -1,0 +1,1 @@
+from .test_eip7441 import *  # noqa: F401 F403
