+++ tests/core/pyspec/eth2spec/test/eip7441/sanity/blocks/test_eip7441.py
@@ -1,8 +1,8 @@
 from eth2spec.test.helpers.block import build_empty_block
-from eth2spec.test.context import spec_state_test, with_whisk_and_later
+from eth2spec.test.context import spec_state_test, with_eip7441_and_later
 from eth2spec.test.helpers.keys import whisk_ks_initial
 from eth2spec.test.helpers.state import state_transition_and_sign_block
-from eth2spec.test.helpers.whisk import compute_whisk_tracker_and_commitment
+from eth2spec.test.helpers.eip7441 import compute_whisk_tracker_and_commitment
 from curdleproofs import WhiskTracker
 
 known_whisk_trackers = {}
@@ -31,7 +31,7 @@ def fill_candidate_trackers(spec, state, tracker: WhiskTracker):
         state.whisk_candidate_trackers[i] = tracker
 
 
-@with_whisk_and_later
+@with_eip7441_and_later
 @spec_state_test
 def test_whisk__process_block_single_initial(spec, state):
     assert state.slot == 0
