+++ tests/core/pyspec/eth2spec/test/eip7441/unittests/test_config_invariants.py
@@ -1,4 +1,4 @@
-from eth2spec.test.context import spec_test, with_whisk_and_later
+from eth2spec.test.context import spec_test, with_eip7441_and_later
 from eth2spec.test.context import single_phase
 
 
@@ -10,7 +10,7 @@ def is_power_of_two(value: int) -> bool:
     return (value > 0) and (value & (value - 1) == 0)
 
 
-@with_whisk_and_later
+@with_eip7441_and_later
 @spec_test
 @single_phase
 def test_curdleproof(spec):
