+++ tests/core/pyspec/eth2spec/test/helpers/constants.py
@@ -17,9 +17,8 @@ ELECTRA = SpecForkName('electra')
 SHARDING = SpecForkName('sharding')
 CUSTODY_GAME = SpecForkName('custody_game')
 DAS = SpecForkName('das')
-ELECTRA = SpecForkName('electra')
 FULU = SpecForkName('fulu')
-WHISK = SpecForkName('whisk')
+EIP7441 = SpecForkName('eip7441')
 EIP7732 = SpecForkName('eip7732')
 
 #
@@ -42,9 +41,9 @@ ALL_PHASES = (
 # The forks that have light client specs
 LIGHT_CLIENT_TESTING_FORKS = (*[item for item in MAINNET_FORKS if item != PHASE0], ELECTRA)
 # The forks that output to the test vectors.
-TESTGEN_FORKS = (*MAINNET_FORKS, ELECTRA, FULU, WHISK)
+TESTGEN_FORKS = (*MAINNET_FORKS, ELECTRA, FULU, EIP7441)
 # Forks allowed in the test runner `--fork` flag, to fail fast in case of typos
-ALLOWED_TEST_RUNNER_FORKS = (*ALL_PHASES, WHISK, EIP7732)
+ALLOWED_TEST_RUNNER_FORKS = (*ALL_PHASES, EIP7441, EIP7732)
 
 # NOTE: the same definition as in `pysetup/md_doc_paths.py`
 PREVIOUS_FORK_OF = {
@@ -56,8 +55,8 @@ PREVIOUS_FORK_OF = {
     DENEB: CAPELLA,
     ELECTRA: DENEB,
     # Experimental patches
-    WHISK: CAPELLA,
     FULU: ELECTRA,
+    EIP7441: CAPELLA,
     EIP7732: ELECTRA,
 }
