+++ tests/core/pyspec/eth2spec/test/helpers/genesis.py
@@ -7,10 +7,10 @@ from eth2spec.test.helpers.execution_payload import (
     compute_el_header_block_hash,
 )
 from eth2spec.test.helpers.forks import (
-    is_post_altair, is_post_bellatrix, is_post_capella, is_post_deneb, is_post_electra, is_post_whisk,
+    is_post_altair, is_post_bellatrix, is_post_capella, is_post_deneb, is_post_electra, is_post_eip7441,
 )
 from eth2spec.test.helpers.keys import pubkeys
-from eth2spec.test.helpers.whisk import compute_whisk_initial_tracker_cached, compute_whisk_initial_k_commitment_cached
+from eth2spec.test.helpers.eip7441 import compute_whisk_initial_tracker_cached, compute_whisk_initial_k_commitment_cached
 
 
 def build_mock_validator(spec, i: int, balance: int):
@@ -153,7 +153,7 @@ def create_genesis_state(spec, validator_balances, activation_threshold):
     if is_post_electra(spec):
         state.deposit_requests_start_index = spec.UNSET_DEPOSIT_REQUESTS_START_INDEX
 
-    if is_post_whisk(spec):
+    if is_post_eip7441(spec):
         vc = len(state.validators)
         for i in range(vc):
             state.whisk_k_commitments.append(compute_whisk_initial_k_commitment_cached(i))
