+++ tests/core/pyspec/eth2spec/test/helpers/specs.py
@@ -3,7 +3,7 @@ from typing import (
 )
 from .constants import (
     MINIMAL, MAINNET,
-    ALL_PHASES, WHISK,
+    ALL_PHASES, EIP7441,
 )
 from .typing import (
     PresetBaseName,
@@ -13,7 +13,7 @@ from .typing import (
 
 
 # NOTE: special case like `ALLOWED_TEST_RUNNER_FORKS`
-ALL_EXECUTABLE_SPEC_NAMES = ALL_PHASES + (WHISK,)
+ALL_EXECUTABLE_SPEC_NAMES = ALL_PHASES + (EIP7441,)
 
 # import the spec for each fork and preset
 for fork in ALL_EXECUTABLE_SPEC_NAMES:
