--- tests/core/pyspec/eth2spec/test/whisk/sanity/blocks/__init__.py
@@ -1,1 +1,0 @@
-from .test_whisk import *  # noqa: F401 F403
